move Signer interface definition and SimpleSigner to the `signer` module
