+++ sdk/src/main/java/me/uport/sdk/signer/MetaIdentitySigner.kt
@@ -1,7 +1,7 @@
 package me.uport.sdk.signer
 
 import me.uport.sdk.MetaIdentityManager
-import me.uport.sdk.core.Signer
+import com.uport.sdk.signer.Signer
 import org.kethereum.extensions.hexToBigInteger
 import org.kethereum.model.Address
 import org.kethereum.model.SignatureData
