+++ signer/build.gradle
@@ -39,6 +39,7 @@ android {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
     implementation "com.android.support:appcompat-v7:$support_lib_version"
 
     api "com.github.walleth.kethereum:crypto:$kethereum_version"
@@ -51,9 +52,16 @@ dependencies {
 
 
 }
-repositories {
-    mavenCentral()
+
+sourceCompatibility = "1.7"
+targetCompatibility = "1.7"
+
+kotlin {
+    experimental {
+        coroutines "enable"
+    }
 }
+
 
 //
 //task androidJavadocs(type: Javadoc) {
